requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
Werkzeug==2.3.7
Gunicorn==21.2.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
Requests==2.31.0
BeautifulSoup4==4.12.2
Selenium==4.15.0
undetected-chromedriver==3.5.3
webdriver-manager==4.0.1
trafilatura==1.6.1
OpenAI==1.3.0
google-genai==0.3.0
SendGrid==6.10.0
email-validator==2.0.0
Flask-Dance==7.0.0
Flask-Login==0.6.2
oauthlib==3.2.2
python-dotenv==1.0.0
APScheduler==3.10.4
Flask-CORS==4.0.0
